WAITM-176 Move getUploadedData into shared
